consist of role based dashboard

+++ frontend/lib/auth.ts
@@ -2,7 +2,7 @@ import api from "./api";
 import { setAuthTokens } from "axios-jwt";
 import { jwtDecode } from "jwt-decode";
 
-export type UserRole = "ADMIN" | "HR" | "EMPLOYEE" | "FINANCE" | "MANAGER";
+export type UserRole = "ADMIN" | "HR" | "EMPLOYEE" | "FINANCE" | "";
 
 export interface JwtPayload {
   id: string;
@@ -22,19 +22,31 @@ export function getUserFromToken(): JwtPayload | null {
 }
 
 export async function login(email: string, password: string) {
-  const res = await api.post("/auth/verify", {
+  const res = await api.post("/auth/login", {
     email,
     password,
   });
-
+  console.log(res)
   setAuthTokens({
     accessToken: res.data.accessToken,
     refreshToken: res.data.refreshToken,
   });
-
   return res.data;
 }
 
-export function logout() {
-  localStorage.clear(); 
+export async function logout() {
+
+  try{
+      const res = await api.post('/auth/logout');
+      console.log(res)
+      if(res.data.success)
+      {
+        localStorage.clear();
+        return res.data
+        
+      }
+  }
+  catch(err){
+    console.log(err);
+  }
 }
